+++ src/models/CharacterFieldNames.java
@@ -1,0 +1,22 @@
+package models;
+
+import java.util.ArrayList;
+
+//public class CharacterFieldNames {
+//	String ageName = "AGE";
+//	String nameName = "NAME";
+//	String genderName = "GENDER";
+//	String alignmentName = "ALIGNMENT";
+//	String speciesName = "SPECIES";
+//	String descriptionName = "DESCRIPTION";
+//	String statName = "STATS";
+//	String strengthName = "STRENGTH";
+//	String perceptionName = "PERCEPTION";
+//	String magicName = "MAGIC POTENTIAL";
+//	String agilityName = "AGILITY";
+//	String willpowerName = "WILLPOWER";
+//	String constitutionName = "CONSTITUTION";
+//	String talentName = "TALENTS";
+//	String techniquesName = "TECHNIQUES";
+//	String otherDescName = "OTHER";
+//}